pyproject.toml
[tool.poetry]
name = "neut"
version = "0.1.0"
description = "A full-stack framework based on FastAPI"
authors = ["Joshua Mayanja <joshua@cybroid.com>"]
readme = "README.md"
homepage = "https://neutpy.com/"
documentation = "https://neutpy.com/docs/"
keywords = ["asgi", "fastapi", "neut"]
classifiers = [
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = "^3.8.1,<4.0"
fastapi = "^0.111.1"
sqlmodel = "^0.0.21"
typer = "^0.12.3"
celery = "^5.4.0"
fastapi-events = "^0.11.1"
firebase-admin = "^6.5.0"
fastapi-mail = "^1.4.1"
python-slugify = "^8.0.4"
pytest = "^8.3.2"
pydantic-settings = "^2.4.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
social-auth-core = "^4.5.4"
pyotp = "^2.9.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
neut = 'neut.core.cli:app'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"